Moves util/workflow.jsx to components/Workflow/WorkflowUtils.jsx and updates imports
